Rename map_metric.py module to map_metric_provider.py module (with DeprecationWarning).
