+++ great_expectations/expectations/metrics/column_map_metrics/column_values_decreasing.py
@@ -11,7 +11,7 @@ from great_expectations.execution_engine.execution_engine import (
     MetricPartialFunctionTypes,
 )
 from great_expectations.expectations.metrics.import_manager import F, Window, sparktypes
-from great_expectations.expectations.metrics.map_metric import (
+from great_expectations.expectations.metrics.map_metric_provider import (
     ColumnMapMetricProvider,
     column_condition_partial,
 )
