+++ great_expectations/expectations/metrics/column_map_metrics/column_values_match_like_pattern.py
@@ -3,7 +3,7 @@ import logging
 from great_expectations.execution_engine.sqlalchemy_execution_engine import (
     SqlAlchemyExecutionEngine,
 )
-from great_expectations.expectations.metrics.map_metric import (
+from great_expectations.expectations.metrics.map_metric_provider import (
     ColumnMapMetricProvider,
     column_condition_partial,
 )
