+++ great_expectations/expectations/metrics/column_map_metrics/column_values_of_type.py
@@ -5,7 +5,7 @@ from great_expectations.execution_engine import PandasExecutionEngine
 from great_expectations.expectations.core.expect_column_values_to_be_of_type import (
     _native_type_type_map,
 )
-from great_expectations.expectations.metrics.map_metric import (
+from great_expectations.expectations.metrics.map_metric_provider import (
     ColumnMapMetricProvider,
     column_condition_partial,
 )
