+++ great_expectations/expectations/metrics/column_map_metrics/column_values_unique.py
@@ -14,12 +14,12 @@ from great_expectations.execution_engine.sqlalchemy_execution_engine import (
     SqlAlchemyExecutionEngine,
 )
 from great_expectations.expectations.metrics.import_manager import F, Window
-from great_expectations.expectations.metrics.map_metric import (
+from great_expectations.expectations.metrics.map_metric_provider import (
     ColumnMapMetricProvider,
     column_condition_partial,
     column_function_partial,
 )
-from great_expectations.expectations.metrics.map_metric import sa as sa
+from great_expectations.expectations.metrics.map_metric_provider import sa as sa
 from great_expectations.expectations.metrics.metric_provider import metric_value
 from great_expectations.validator.validation_graph import MetricConfiguration
