+++ great_expectations/expectations/metrics/column_pair_map_metrics/column_pair_values_greater.py
@@ -7,7 +7,9 @@ from great_expectations.execution_engine.execution_engine import (
     MetricDomainTypes,
     MetricPartialFunctionTypes,
 )
-from great_expectations.expectations.metrics.map_metric import MapMetricProvider
+from great_expectations.expectations.metrics.map_metric_provider import (
+    MapMetricProvider,
+)
 from great_expectations.expectations.metrics.metric_provider import metric_partial
 from great_expectations.expectations.metrics.util import filter_pair_metric_nulls
